删除了一些工具类和依赖包，修改了下代码

+++ model/type.go
@@ -1,7 +1,7 @@
 package model
 
 type Type struct {
-	Id       uint8  `json:"id"`
+	Id       uint16 `json:"id"`
 	TypeName string `json:"typeName" gorm:"size:10;NOT NULL"`
 	Amount   uint32 `json:"amount"`
 	Blogs    []Blog
